+++ src/main/java/org/aes/searchnode/business/concretes/searchnode/SearchNode.java
@@ -55,7 +55,7 @@ public class SearchNode<T> implements SearchNodeService<T> {
     }
 
     @Override
-    public long getTotalItemNumber() {
+    public int getTotalItemNumber() {
         return searchNodeService.getTotalItemNumber();
     }
